+++ scratch.js
@@ -1,6 +1,7 @@
 const nlp = require('./src/index')
-nlp.extend(require('./plugins/numbers/src'))
-nlp.extend(require('./plugins/dates/src'))
+// nlp.extend(require('./plugins/numbers/src'))
+// nlp.extend(require('./plugins/dates/src'))
+nlp.extend(require('./plugins/sentences/src'))
 // nlp.verbose(true)
 
 //
@@ -13,8 +14,21 @@ nlp.extend(require('./plugins/dates/src'))
 // complex denominators - 'one fifty fourths', 'one thirty third'
 //
 
-// let doc = nlp(`30mins tuesday`).debug()
-let doc = nlp(`for 20 mins`).debug()
-// let doc = nlp(`in 20 mins`).debug()
-console.log(doc.dates().get())
+// let doc = nlp('i should drive')
+// doc.sentences().toPastTense().debug()
+
+// let doc = nlp(`shut the door`).debug()
+// let doc = nlp(`do you eat?`) //.debug()
+let doc = nlp(`the service is fast`).debug()
+
+// let vb = doc.verbs().clone(true)
+// vb.sentences().debug()
+// doc.verbs().isImperative().debug()
+// doc.sentences().toPastTense().debug()
+
+// console.log(doc.dates().get())
 // console.log(doc.durations().get(0))
+
+// possible match-bug:
+// let doc = nlp(`go fast john!`).debug()
+// s.has('^#Infinitive #Adverb? #Noun?$')